+++ src/creators/createFolder.ts
@@ -1,34 +1,29 @@
-import { textRetry } from '../constants';
 import { join } from 'path';
-import { window } from 'vscode';
-import { CommCreateComponent, removeDot } from '../helpers';
 import { access, mkdir } from 'fs/promises';
+import { finishProcess } from '../helpers/finish-process';
+import { NameComponent } from '../showInput/InputNameComponent';
 
-const createFolder = async (path: string, componentName: string) => {
+export const createFolder = async (
+	path: string,
+	componentName: NameComponent
+): Promise<string> => {
 	// Crear una carpeta con el nombre del componente
-	const nameFolder = removeDot(componentName);
-
-	const folderPath = join(path, nameFolder);
+	const folderPath = join(path, componentName.capitalize);
 
 	// Verifica si la carpeta ya existe
 	try {
 		await access(folderPath);
-		window
-			.showErrorMessage(
-				`The Component "${componentName}" already exists in the path ${path}`,
-				textRetry
-			)
-			.then(selection => selection === textRetry && CommCreateComponent());
-		return null;
+
+		finishProcess(
+			`The Component "${componentName.capitalize}" already exists in the path ${folderPath}`
+		);
 	} catch (err) {}
 
 	try {
 		await mkdir(folderPath, { recursive: true });
 	} catch (err: any) {
-		return window.showErrorMessage(err.message);
+		finishProcess(err.message);
 	}
 
 	return folderPath;
 };
-
-export { createFolder };