+++ src/extension.ts
@@ -1,59 +1,66 @@
 import { ExtensionContext, commands, window } from 'vscode';
-import { removeDot, showMessageCancel } from './helpers';
+import { ExtensionStyle, removeDot } from './helpers';
 import { comandoCreateComponent, NOT_STYLE } from './constants';
 
 import {
-	checkPath,
-	InputNameComponent,
-	selectLanguage,
+	getLanguage,
+	getNameComponent,
+	getPath,
+	getStyleType,
 	selectStyleLanguage,
-	selectStyleType,
 } from './showInput';
 
 import { createFolder, createStyles, createComponent, createBarrer } from './creators';
 
-export function activate(context: ExtensionContext) {
+export async function activate(context: ExtensionContext) {
 	let disposable = commands.registerCommand(comandoCreateComponent, async args => {
-		// Valida la ruta del proyecto
-		const path = await checkPath(args);
-		if (!path) return showMessageCancel();
+		// Obtiene la ruta del archivo seleccionado
+		const path = args?.fsPath ?? (await getPath(context));
 
 		// Selecciona el lenguaje
-		const COMPONENT_STRUCTURE = await selectLanguage();
-		if (!COMPONENT_STRUCTURE) return showMessageCancel();
+		const language = await getLanguage();
 
 		// El tipo de diseño que tendrá el componente, ej: style-component
-		let STYLE_TYPE = await selectStyleType();
-		if (!STYLE_TYPE) return showMessageCancel();
+		const styleType = await getStyleType();
 
-		// Selecciona el lenguaje de la hoja de estilos
-		const STYLE_LANGUAGE = await selectStyleLanguage(STYLE_TYPE);
-		if (!STYLE_LANGUAGE) return showMessageCancel();
+		// Selecciona la extension de la hoja de estilos
+		const extensionStyle =
+			styleType === 'Style Component' || styleType === 'Not style'
+				? NOT_STYLE
+				: await selectStyleLanguage();
 
 		// Pregunta por el nombre del componente
-		const COMPONENT_NAME = await InputNameComponent();
-		if (!COMPONENT_NAME) return showMessageCancel();
+		const nameComponent = await getNameComponent();
 
-		const FOLDER_PATH = await createFolder(path, COMPONENT_NAME);
-		if (!FOLDER_PATH) return;
+		const folderPath = await createFolder(path, nameComponent);
 
-		if (STYLE_LANGUAGE !== NOT_STYLE)
-			createStyles(FOLDER_PATH, COMPONENT_NAME, STYLE_TYPE, STYLE_LANGUAGE);
+		if (extensionStyle !== NOT_STYLE) {
+			createStyles(folderPath, nameComponent, styleType, extensionStyle);
+		}
 
 		createComponent(
-			FOLDER_PATH,
-			COMPONENT_NAME,
-			COMPONENT_STRUCTURE,
-			STYLE_TYPE,
-			STYLE_LANGUAGE
+			folderPath,
+			nameComponent,
+			language,
+			styleType,
+			extensionStyle as ExtensionStyle
 		);
 
-		createBarrer(COMPONENT_NAME, FOLDER_PATH, COMPONENT_STRUCTURE);
+		createBarrer(nameComponent, folderPath, language);
 
-		window.showInformationMessage(`Component ${removeDot(COMPONENT_NAME)} created. 🎉`);
+		window.showInformationMessage(
+			`Component ${removeDot(nameComponent.capitalize)} created. ⚛`
+		);
 	});
 
-	context.subscriptions.push(disposable);
+	const deleteCache = commands.registerCommand(
+		'Create-component-React:delete-cache',
+		async args => {
+			context.workspaceState.update('routes', []);
+		}
+	);
+
+	context.subscriptions.push(disposable, deleteCache);
 }
 
 export function deactivate() {}