+++ src/helpers/ShowMessageCancel.ts
@@ -1,9 +1,9 @@
 import { window, commands } from 'vscode';
 import { comandoCreateComponent, messageCancel, textRetry } from '../constants';
 
-export const showMessageCancel = () =>
+export const showMessageCancel = (message: string = messageCancel) =>
 	window
-		.showErrorMessage(messageCancel, textRetry)
+		.showErrorMessage(message, textRetry)
 		.then(selection => selection === textRetry && CommCreateComponent());
 
 export const CommCreateComponent = () => commands.executeCommand(comandoCreateComponent);