+++ src/templates/barrel.template.ts
@@ -1,2 +1,2 @@
-export const templateBarrel = (COMPONENT_NAME: string, COMPONENT_NAME_FILE: string) =>
-	`export { default as ${COMPONENT_NAME} } from './${COMPONENT_NAME_FILE}';`;
+export const templateBarrel = (nameComponent: string) =>
+	`export { default as ${nameComponent} } from './${nameComponent}';`;