+++ CHANGELOG.md
@@ -1,9 +1,59 @@
-# Changelog
+# CHANGELOG
 
-| Version | Date       | Changelog                                                 |
-| ------- | ---------- | --------------------------------------------------------- |
-| 1.0.1   | 26.09.2022 | Changelog                                                 |
-| 1.0.0   | 26.09.2022 | bug fix and completion of component creation for react js |
-| 0.0.4   | 30.08.2022 | aggregation of settings by the workspace                  |
-| 0.0.3   | 28.08.2022 | update readme                                             |
-| 0.0.1   | 27.08.2022 | extension creation                                        |
+## 1.1.0 (14 abril 2023)
+
+### Search path
+
+- Added a complemented car to search for routes
+- Added a `Recommended Options` section that can be added or removed by your `WorkSpace`
+- Added a `Recent Options` section to only select and not search for your route again
+
+### Typescript
+
+- Added types that the project did not have
+
+### Select
+
+- All the select are added more information and details in their options for their use
+
+---
+
+## 1.0.4 ( 26.09.2022)
+
+- We remove the default `interface` name and change it to `props`
+
+---
+
+## 1.0.2 ( 26.09.2022)
+
+- Added auto check extension functionality
+
+---
+
+## 1.0.1 ( 26.09.2022)
+
+- Changelog
+
+---
+
+## 1.0.0 ( 26.09.2022)
+
+- Bug fix and completion of component creation for react js
+
+---
+
+## 0.0.4 (30.08.2022)
+
+- Aggregation of settings by the workspace |
+
+---
+
+## 0.0.3 (28 agosto 2022)
+
+- Update readme
+
+---
+
+## 0.0.1 (27 agosto 2022)
+
+- Extension creation